Move files, use uuidv4, and clonedeep the state
